ESS-124-automatedTest #comment Reserve parameter for bot browser

+++ test-client.sh
@@ -1,12 +1,24 @@
 #!/bin/sh
 
+test="$1";
+browser="$2";
+
 # karma should only run a single test suite here
 # Temporary solution: use different karma config files
 # Possible to accept argument to karma and load only specific files ?
-# TODO: Replace chrome with parameter
-karma start karma-config/"$2"/karma-"$1".conf.js
+karma start karma-config/$browser/karma-$test.conf.js
 
 #After test was run, close all remaining bots and node processes
-#TODO: close browser based on parameter
-killall -9 "Google Chrome"
+
+kill_browser(){
+	if [ $browser == "chrome" ]; 
+	then
+		killall -9 "Google Chrome"
+	elif [ $browser == "firefox" ];
+	then
+		killall -9 "Firefox"
+	fi
+}
+
+kill_browser
 killall -9 node
